requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
argon2-cffi=20.1.0=py38he774522_1
async_generator=1.10=py_0
attrs=20.2.0=py_0
backcall=0.2.0=py_0
blas=1.0=mkl
bleach=3.2.1=py_0
ca-certificates=2020.10.14=0
cachetools=4.1.1=pypi_0
certifi=2020.6.20=pyhd3eb1b0_3
cffi=1.14.3=py38h7a1dbc1_0
chardet=3.0.4=pypi_0
colorama=0.4.4=py_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
entrypoints=0.3=py38_0
google-api-core=1.23.0=pypi_0
google-api-python-client=1.12.5=pypi_0
google-auth=1.23.0=pypi_0
google-auth-httplib2=0.0.4=pypi_0
google-auth-oauthlib=0.4.2=pypi_0
googleapis-common-protos=1.52.0=pypi_0
gspread=3.6.0=pypi_0
httplib2=0.18.1=pypi_0
icu=58.2=ha925a31_3
idna=2.10=pypi_0
importlib-metadata=2.0.0=py_1
importlib_metadata=2.0.0=1
intel-openmp=2020.2=254
ipykernel=5.3.4=py38h5ca1d4c_0
ipython=7.18.1=py38h5ca1d4c_0
ipython_genutils=0.2.0=py38_0
ipywidgets=7.5.1=py_1
jedi=0.17.2=py38_0
jinja2=2.11.2=py_0
jpeg=9b=hb83a4c4_2
jsonschema=3.2.0=py_2
jupyter=1.0.0=py38_7
jupyter_client=6.1.7=py_0
jupyter_console=6.2.0=py_0
jupyter_core=4.6.3=py38_0
jupyterlab_pygments=0.1.2=py_0
libpng=1.6.37=h2a8f88b_0
libsodium=1.0.18=h62dcd97_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=1.1.1=py38he774522_0
mistune=0.8.4=py38he774522_1000
mkl=2020.2=256
mkl-service=2.3.0=py38hb782905_0
mkl_fft=1.2.0=py38h45dec08_0
mkl_random=1.1.1=py38h47e9c7a_0
msys2-conda-epoch=20160418=1
nbclient=0.5.1=py_0
nbconvert=6.0.7=py38_0
nbformat=5.0.8=py_0
nest-asyncio=1.4.1=py_0
notebook=6.1.4=py38_0
numpy=1.19.2=py38hadc3359_0
numpy-base=1.19.2=py38ha3acd2a_0
oauth2client=4.1.3=pypi_0
oauthlib=3.1.0=pypi_0
openssl=1.1.1h=he774522_0
packaging=20.4=py_0
pandas=1.1.3=py38ha925a31_0
pandoc=2.11=h9490d1a_0
pandocfilters=1.4.2=py38_1
parso=0.7.0=py_0
pickleshare=0.7.5=py38_1000
pip=20.2.4=py38haa95532_0
prometheus_client=0.8.0=py_0
prompt-toolkit=3.0.8=py_0
prompt_toolkit=3.0.8=0
protobuf=3.13.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.20=py_2
pygments=2.7.1=py_0
pyparsing=2.4.7=py_0
pyqt=5.9.2=py38ha925a31_4
pyrsistent=0.17.3=py38he774522_0
python=3.8.5=h5fd99cc_1
python-dateutil=2.8.1=py_0
pytz=2020.4=pypi_0
pywin32=227=py38he774522_1
pywinpty=0.5.7=py38_0
pyzmq=19.0.2=py38ha925a31_1
qt=5.9.7=vc14h73c81de_0
qtconsole=4.7.7=py_0
qtpy=1.9.0=py_0
requests=2.24.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.6=pypi_0
send2trash=1.5.0=py38_0
setuptools=50.3.0=py38haa95532_1
sip=4.19.13=py38ha925a31_0
six=1.15.0=py_0
sqlite=3.33.0=h2a8f88b_0
terminado=0.9.1=py38_0
testpath=0.4.4=py_0
tornado=6.0.4=py38he774522_1
traitlets=5.0.5=py_0
uritemplate=3.0.1=pypi_0
urllib3=1.25.11=pypi_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_3
wcwidth=0.2.5=py_0
webencodings=0.5.1=py38_1
wheel=0.35.1=py_0
widgetsnbextension=3.5.1=py38_0
wincertstore=0.2=py38_0
winpty=0.4.3=4
zeromq=4.3.2=ha925a31_3
zipp=3.4.0=pyhd3eb1b0_0
zlib=1.2.11=h62dcd97_4
